scripts/tts/requirements.txt
flask==3.0.0
azure-cognitiveservices-speech==1.34.0
supabase==2.0.2
pydub==0.25.1
python-dotenv==1.0.0